SG-iCalendar: Fix parsing of invalid timezones created by Outlook iCals.
Outlook iCals use the following format for their dates:

    DTSTART;TZID="(UTC-05:00) Eastern Time (US & Canada)":20160921T090000

There are two problems with this.

1. TZID="(UTC-05:00) Eastern Time (US & Canada)" is an invalid iCal timezone.
2. SG-iCalendar doesn't like colons used twice on one line as it interferes
   with how they parse timezones and the date timestamp.

To remedy this, we just strip out the entire TZID="()" string so
SG-iCalendar can continue parsing events properly.

In the future, we could do something fancy like mapping the invalid
timezone to the proper iCal timezone.

Fixes #18.