SG-iCalendar: Fix parsing of Microsoft timezones created by Outlook iCals.
Previously, in commit 7e8478e, we stripped the TZID from event times.  This
created problems when parsing and creating events with a custom timezone.

In order to fix the parsing of TZIDs in SG-iCalendar, we're seeing if the
event time is using a Microsoft timezone.  If so, we try to match the
Microsoft timezone (eg. "(UTC-05:00) Eastern Time (US & Canada)") to an
Olson timezone (eg. America/Toronto).  If there is a match, we convert the
event time to UTC.

Fixes #19.